still on inheritence from the bp first person character
Going to look into defining a higher level first person character that acts as the parent to bp first person character. Need to learn how to set up the input mapping context as functions that can be changed in children instances of the player (as in when we jump cameras and lock onto objects).