Basis of desk peak is working
Implemented a hierarchy class system of cameras, as well as a spawner that can spawn cameras in as soon as we interact with objects that will need them. This will also dynamically set the camera position and rotation (including rotation restraints) to the relative position and rotation of the player controller in the game.

still need to do the following:
- destroy generated cameras as soon as we are done using them
- create cameras for different interaction levels (we currently have one that is meant for camera rotation only)